Test Unitario Crear Cliente

+++ BLL/Vendedor.cs
@@ -18,7 +18,7 @@ namespace BLL
             set { _CantidadVentas = value; }
         }
 
-        public bool RegistrarCliente(string idCliente, string nombre, string apellido, int dni, int cantidadCompras)
+        public bool RegistrarCliente(string idCliente, string nombre, string apellido, long dni, long cantidadCompras)
         {
             DAL.Vendedor cliente = new DAL.Vendedor();
